Split models module into functional and image models
