+++ photutils/psf/__init__.py
@@ -4,14 +4,15 @@ This subpackage contains tools to perform point-spread-function (PSF)
 photometry.
 """
 
-from . import (epsf, epsf_stars, griddedpsfmodel, groupers, models,
-               photometry, utils)
+from . import (epsf, epsf_stars, functional_models, griddedpsfmodel, groupers,
+               image_models, photometry, utils)
 from .epsf import *  # noqa: F401, F403
 from .epsf_stars import *  # noqa: F401, F403
+from .functional_models import *  # noqa: F401, F403
 from .griddedpsfmodel import *  # noqa: F401, F403
 from .groupers import *  # noqa: F401, F403
+from .image_models import *  # noqa: F401, F403
 from .matching import *  # noqa: F401, F403
-from .models import *  # noqa: F401, F403
 from .photometry import *  # noqa: F401, F403
 from .utils import *  # noqa: F401, F403
 
@@ -19,8 +20,9 @@ from .utils import *  # noqa: F401, F403
 __all__ = []
 __all__ += epsf.__all__
 __all__ += epsf_stars.__all__
+__all__ += functional_models.__all__
 __all__ += griddedpsfmodel.__all__
 __all__ += groupers.__all__
-__all__ += models.__all__
+__all__ += image_models.__all__
 __all__ += photometry.__all__
 __all__ += utils.__all__
