+++ photutils/psf/epsf.py
@@ -16,7 +16,7 @@ from astropy.utils.exceptions import AstropyUserWarning
 
 from photutils.centroids import centroid_com
 from photutils.psf.epsf_stars import EPSFStar, EPSFStars, LinkedEPSFStar
-from photutils.psf.models import EPSFModel
+from photutils.psf.image_models import EPSFModel
 from photutils.psf.utils import _interpolate_missing_data
 from photutils.utils._optional_deps import HAS_BOTTLENECK
 from photutils.utils._parameters import as_pair
